+++ vortex-array/src/compute/compare.rs
@@ -148,13 +148,13 @@ impl ComputeFnVTable for Compare {
         }
 
         // First try lhs op rhs, then invert and try again.
-        if let Some(output) = lhs.invoke(&COMPARE_FN, args)? {
-            return Ok(output);
-        }
         for kernel in kernels {
             if let Some(output) = kernel.invoke(args)? {
                 return Ok(output);
             }
+        }
+        if let Some(output) = lhs.invoke(&COMPARE_FN, args)? {
+            return Ok(output);
         }
 
         // Try inverting the operator and swapping the arguments
@@ -162,13 +162,13 @@ impl ComputeFnVTable for Compare {
             inputs: &[rhs.into(), lhs.into()],
             options: &operator.swap(),
         };
-        if let Some(output) = rhs.invoke(&COMPARE_FN, &inverted_args)? {
-            return Ok(output);
-        }
         for kernel in kernels {
             if let Some(output) = kernel.invoke(&inverted_args)? {
                 return Ok(output);
             }
+        }
+        if let Some(output) = rhs.invoke(&COMPARE_FN, &inverted_args)? {
+            return Ok(output);
         }
 
         // Only log missing compare implementation if there's possibly better one than arrow,
